[#457] make getCity hook reusable
