+++ frontend/app/[city]/kap_checkliste/page.tsx
@@ -1,0 +1,9 @@
+import { Container } from "react-bootstrap";
+
+export default function CapChecklist() {
+  return (
+    <Container>
+      <h2>Klimaaktionsplan </h2>
+    </Container>
+  );
+}
