+++ frontend/app/[city]/page.tsx
@@ -1,12 +1,12 @@
 "use client";
 
-import axios from "axios";
 import Image from "next/image";
 import { usePathname } from "next/navigation";
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { Container } from "react-bootstrap";
 import Markdown from "react-markdown";
 import arrow from "../../public/images/arrow-right-down.svg";
+import { useGetCity } from "../CityService";
 import LocalGroup from "../components/LocalGroup";
 
 const CityDescription = ({ description }) => {
@@ -34,40 +34,18 @@ const SupportingNgos = ({ supportingNgos }) => {
 };
 
 export default function CityDashboard() {
-  const [city, setCity] = useState({});
-  const [hasError, setHasError] = useState(false);
   const [isLocalGroupExpanded, setIsLocalGroupExpanded] = useState(false);
 
   const pathname = usePathname();
   const slug = pathname.split("/").at(-1);
 
-  const getCity = async () => {
-    const response = await axios
-      .get("http://127.0.0.1:8000/api/cities/" + slug, {
-        // TODO: proper url
-        headers: {
-          Accept: "application/json",
-          "Content-Type": "application/json;charset=UTF-8",
-        },
-      })
-      .then((response) => {
-        setCity(response.data);
-      })
-      .catch((error) => {
-        setHasError(true);
-        console.error("Error get city request: ", error);
-      });
-  };
-
-  useEffect(() => {
-    getCity();
-  }, []);
+  const { city, hasError } = useGetCity(slug);
 
   if (hasError) {
     return <h3 className="pb-3 pt-3">Für die Stadt {slug} gibt es kein Monitoring</h3>;
   }
 
-  if (!city.name) {
+  if (!city) {
     return <></>;
   }
